+++ packages/react-components/react-aria/src/index.ts
@@ -1,2 +1,3 @@
-export * from './hooks/index';
-export * from './utils/index';
+export { useARIAButton } from './hooks/index';
+export type { ARIAButtonSlotProps } from './hooks/index';
+export { mergeARIADisabled } from './utils/index';