+++ packages/react-components/react-checkbox/src/components/Checkbox/useCheckboxStyles.ts
@@ -15,10 +15,6 @@ export const checkboxClassNames: SlotClassNames<CheckboxSlots> = {
   indicator: 'fui-Checkbox__indicator',
 };
 
-// TODO replace these spacing constants with theme values once they're on the theme
-const spacingHorizontalS = '8px';
-const spacingHorizontalM = '12px';
-
 // The indicator size is used by the indicator and label styles
 const indicatorSizeMedium = '16px';
 const indicatorSizeLarge = '20px';
@@ -27,8 +23,7 @@ const useRootStyles = makeStyles({
   base: {
     position: 'relative',
     display: 'inline-flex',
-    columnGap: spacingHorizontalM,
-    ...shorthands.padding(spacingHorizontalS),
+    ...shorthands.padding(tokens.spacingVerticalS, tokens.spacingHorizontalS),
     ...createFocusOutlineStyle({ style: {}, selector: 'focus-within' }),
   },
 });
@@ -196,13 +191,22 @@ const useLabelStyles = makeStyles({
     color: 'inherit',
   },
 
+  before: {
+    marginRight: tokens.spacingHorizontalM,
+  },
+  after: {
+    marginLeft: tokens.spacingHorizontalM,
+  },
+
   // Use a (negative) margin to account for the difference between the indicator's height and the label's line height.
   // This prevents the label from expanding the height of the Checkbox, but preserves line height if the label wraps.
   medium: {
-    ...shorthands.margin(`calc((${indicatorSizeMedium} - ${tokens.lineHeightBase300}) / 2)`, 0),
+    marginTop: `calc((${indicatorSizeMedium} - ${tokens.lineHeightBase300}) / 2)`,
+    marginBottom: `calc((${indicatorSizeMedium} - ${tokens.lineHeightBase300}) / 2)`,
   },
   large: {
-    ...shorthands.margin(`calc((${indicatorSizeLarge} - ${tokens.lineHeightBase300}) / 2)`, 0),
+    marginTop: `calc((${indicatorSizeLarge} - ${tokens.lineHeightBase300}) / 2)`,
+    marginBottom: `calc((${indicatorSizeLarge} - ${tokens.lineHeightBase300}) / 2)`,
   },
 });
 
@@ -233,6 +237,7 @@ export const useCheckboxStyles_unstable = (state: CheckboxState): CheckboxState 
       checkboxClassNames.label,
       labelStyles.base,
       labelStyles[state.size],
+      labelStyles[state.labelPosition],
       state.label.className,
     );
   }