+++ packages/react-components/react-positioning/src/index.ts
@@ -1,6 +1,21 @@
-export * from './createVirtualElementFromClick';
-export * from './createArrowStyles';
-export * from './usePopper';
-export * from './usePopperMouseTarget';
+export { createVirtualElementFromClick } from './createVirtualElementFromClick';
+export { createArrowHeightStyles, createArrowStyles } from './createArrowStyles';
+export type { CreateArrowStylesOptions } from './createArrowStyles';
+export { usePopper } from './usePopper';
+export { usePopperMouseTarget } from './usePopperMouseTarget';
 export { resolvePositioningShorthand, mergeArrowOffset } from './utils/index';
-export * from './types';
+export type {
+  Alignment,
+  AutoSize,
+  Boundary,
+  Offset,
+  OffsetFunction,
+  OffsetFunctionParam,
+  PopperOptions,
+  PopperRefHandle,
+  PopperVirtualElement,
+  Position,
+  PositioningProps,
+  PositioningShorthand,
+  PositioningShorthandValue,
+} from './types';