+++ packages/react-components/react-utilities/CHANGELOG.md
@@ -1,8 +1,18 @@
 # Change Log - @fluentui/react-utilities
 
-This log was last generated on Wed, 04 May 2022 13:26:36 GMT and should not be manually modified.
+This log was last generated on Thu, 05 May 2022 18:26:28 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [9.0.0-rc.8](https://github.com/microsoft/fluentui/tree/@fluentui/react-utilities_v9.0.0-rc.8)
+
+Thu, 05 May 2022 18:26:28 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-utilities_v9.0.0-rc.7..@fluentui/react-utilities_v9.0.0-rc.8)
+
+### Changes
+
+- feat: add time element on nativeElementMap ([PR #22848](https://github.com/microsoft/fluentui/pull/22848) by bsunderhus@microsoft.com)
+- Bump @fluentui/keyboard-keys to v9.0.0-rc.6 ([PR #22857](https://github.com/microsoft/fluentui/pull/22857) by beachball)
 
 ## [9.0.0-rc.7](https://github.com/microsoft/fluentui/tree/@fluentui/react-utilities_v9.0.0-rc.7)